Tidying C interface.

+++ Wrappers/C/CubiquityC.h
@@ -1,3 +1,5 @@
+#include <stdint.h> //C99 fixed size data types.
+
 // The following ifdef block is the standard way of creating macros which make exporting 
 // from a DLL simpler. All files within this DLL are compiled with the CUBIQUITYC_EXPORTS
 // symbol defined on the command line. This symbol should not be defined on any project
@@ -13,16 +15,16 @@
 extern "C"
 {
 	// Volume functions
-	CUBIQUITYC_API unsigned int newColouredCubesVolume(int lowerX, int lowerY, int lowerZ, int upperX, int upperY, int upperZ, unsigned int blockSize, unsigned int baseNodeSize);
-	CUBIQUITYC_API void updateVolume(unsigned int volumeHandle);
+	CUBIQUITYC_API unsigned int newColouredCubesVolume(int32_t lowerX, int32_t lowerY, int32_t lowerZ, int32_t upperX, int32_t upperY, int32_t upperZ, uint32_t blockSize, uint32_t baseNodeSize);
+	CUBIQUITYC_API void updateVolume(uint16_t volumeHandle);
 
 	// Octree functions
-	CUBIQUITYC_API unsigned int getRootOctreeNode(unsigned int volumeHandle);
+	CUBIQUITYC_API uint16_t getRootOctreeNode(uint16_t volumeHandle);
 
 	// Mesh functions
-	CUBIQUITYC_API unsigned int getNoOfVertices(unsigned int volumeHandle, unsigned int octreeNodeHandle);
-	CUBIQUITYC_API unsigned int getNoOfIndices(unsigned int volumeHandle, unsigned int octreeNodeHandle);
+	CUBIQUITYC_API unsigned int getNoOfVertices(uint16_t volumeHandle, uint16_t octreeNodeHandle);
+	CUBIQUITYC_API unsigned int getNoOfIndices(uint16_t volumeHandle, uint16_t octreeNodeHandle);
 
-	CUBIQUITYC_API float* getVertices(unsigned int volumeHandle, unsigned int octreeNodeHandle);
-	CUBIQUITYC_API unsigned int* getIndices(unsigned int volumeHandle, unsigned int octreeNodeHandle);
+	CUBIQUITYC_API float* getVertices(uint16_t volumeHandle, uint16_t octreeNodeHandle);
+	CUBIQUITYC_API unsigned int* getIndices(uint16_t volumeHandle, uint16_t octreeNodeHandle);
 }
